refactor: :recycle: refactor vehicle control code into it's own class/file (with helper functions)
